+++ Test/TestMain.fs
@@ -1,6 +1,10 @@
-﻿module Euclid.BVH.Tests
+module Euclid.BVH.Tests
 open System
 
+#if FABLE_COMPILER_JAVASCRIPT || FABLE_COMPILER_TYPESCRIPT
+open Fable.Mocha
+let test x = Mocha.runTests x
+#else
 open Expecto
 open System.Globalization
 open System.Threading
@@ -8,11 +12,17 @@ Thread.CurrentThread.CurrentCulture   <- CultureInfo.GetCultureInfo "en-US" // s
 Thread.CurrentThread.CurrentUICulture <- CultureInfo.GetCultureInfo "en-US"
 let mutable cliArgs : string[] = [||]
 let test x = runTestsWithCLIArgs [] cliArgs x
+#endif
 
 let run () =
     test TestBvh.tests
     |||
     test TestLineBvh.tests
+
+#if FABLE_COMPILER_JAVASCRIPT || FABLE_COMPILER_TYPESCRIPT
+#nowarn "20" //The result of this expression has type 'int' and is implicitly ignored.
+run()
+#else
 
 [<EntryPoint>]
 let main (args: string[]) =
@@ -23,3 +33,4 @@ let main (args: string[]) =
     else
         printfn "%d tests failed" r
     r
+#endif